Starting to add message encoding for zebra
I think that the in and out of the zebra socket can be opened
separately, and I can send a message to the out while a thread is
waiting on the in. I have to test this. In the meantime, I created a
strategy pattern of sorts, with the addition of the encode method to the
routing messages, that would encode the message to send to the zebra
socket.