Async message sent!
Tested a RabbitMQ async message. Seems to work. We'll decide whether to
use the already existing RabbitMQ infrastructure, or ActiveMQ. RabbitMQ
has to be installed and run separately, while ActiveMQ is a pure Java
solution that can be started from within the server. We'll design an
architecture contemplating both cases.