Adding a zebra socket
I just got tired of searching for a way to communicate with Unix sockets
using asynchronous methods. If what I can do is using a thread that will
be ultimately blocked while waiting, I'll do that. I'm going to use a
channel to communicate from there with the rest of the server, anyway,
so the communication will be guaranteed to be thread-safe, or at least,
I'll try. I may have to make the method synchronized.