Reforms
I reached the conclusion that I have to separate the messages sent
through the Zebra socket from those that the Zebra daemon is able to
receive. They may be the same, but the coding/encoding functions are
obviously different, since the RoutingMessage can be created by the
BgpHandler and simply serialized through the RabbitMQ channel, and the
RoutingMessage has to be encoded to be sent through the socket.